Added autogen.sh: just run ./autogen.sh to generate a new configure file after having modified a Makefile.am or configure.ac
The configure file in CVS is now up to date

Patched a bug in mpfi_trigo.c affecting mpfi_sin when called on a symetrical
interval

Patched an old bug in mpfi_trigo.c in function mpfi_mpz_exact_set_fr
This bug had already been patched once since the RC available as a tgz on mpfi's
web page is correct.
It is quite surprising to find this bug in the CVS...

Corrected the version number in configure.ac

Patched configure.ac and some  Makefile.am in order to use libtool to
generate a dynamic library.

Warning: the libraries are now compiled in src/.libs/


git-svn-id: svn://scm.gforge.inria.fr/svnroot/mpfi@11 e2f78e0c-bb60-4709-9413-9660a31d4696